Introduce ElectricityMeter as replacement for SymmetricMeter & AsymmetricMeter (#2144)
Distinguishing between `SymmetricMeter` and `AsymmetricMeter` and having `AsymmetricMeter extends SymmetricMeter` did never really make sense physically. In fact the _asymmetric_ meter (i.e. the three-phase meter) is the _most normal_ case. Special cases are:
- a _symmetric_ meter, i.e. one that has the same values on all phases; this can be handled entirely by deriving the individual phase values from the sum values, via dividing the power by three.
- a _single-phase_ meter, i.e. one that only has values on one phase. This requires active configuration and then allows setting the other phases to zero. The `SinglePhaseMeter` Nature is kept, because it allows a use-case when UI should really only show one value for these devices.

The special cases can be represented properly by this newly introduced `ElectricityMeter`, which combines most of the Channels and replaces SymmetricMeter and AsymmetricMeter entirely. It also clearly distinguishes _electricity meters_ from _heat meters_ that are slowly coming to OpenEMS.

This PR includes:
- Replace SymmetricMeter and AsymmetricMeter
- Breaking:
  - Renamed Meter.Virtual.Asymmetric.Add -> Meter.Virtual.Add
  - Dropped Meter.Virtual.Symmetric.Add
  - Renamed Meter.Virtual.Symmetric.Subtract -> Meter.Virtual.Subtract
- Simulator Grid Meter Reacting: add Voltage and Current sum and per phase for better testing
- UI fixes
  - show negative current per phase
  - fix translation issues

It also tries to clarify the sometimes confusing representation of positive and negative values for Power and Current. See the Javadoc in ElectricityMeter and the discussion at https://community.openems.io/t/change-energy-channel-assignment-based-on-meter-type/1603/6.

As this PR touches a lot of files, it is kept as small, simple and non-breaking as possible. Follow-up PRs will add more features,